requirements.txt
google-cloud-bigquery
google-cloud-core  
google-auth  
google.resumable.media  
google.api.core
google-auth-oauthlib
pandas
pandas-gbq
db-dtypes
plotly
gcsfs
st-files-connection
gcloud auth login
gcloud auth application-default login